+++ src/mastodon_api/subscriptions/types.rs
@@ -38,6 +38,7 @@ impl From<DbInvoice> for Invoice {
             InvoiceStatus::Forwarded => "forwarded",
             InvoiceStatus::Timeout => "timeout",
             InvoiceStatus::Cancelled => "cancelled",
+            InvoiceStatus::Underpaid => "underpaid",
         };
         let expires_at = if value.chain_id.inner().is_monero() {
             value.created_at + Duration::seconds(MONERO_INVOICE_TIMEOUT)