+++ CHANGELOG.md
@@ -27,6 +27,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Rename `check-expired-invoice` command to `reopen-invoice` and allow to reopen invoices with "forwarded" status.
 - Allow to call `reopen-invoice` command with payment address as an argument.
 - Verify that account index returned by monero-wallet-rpc matches configuration.
+- Change invoice status to "underpaid" if amount is too small to be forwarded.
 
 ### Deprecated